ENH + TEST: enforce float image affines
Image affines could be any data type (depending on what was passed).
This could be odd when doing things like ``img.get_affine()[0, 0] = 1.1;
the value would be truncated if the original affine was an integer type.
Obviously this is numpy's default behavior, but we're copying the affine
anyway, so casting to float seems ordinary and a bit less surprising.

+++ nibabel/spatialimages.py
@@ -301,8 +301,9 @@ class SpatialImage(object):
             # Check that affine is array-like 4,4.  Maybe this is too strict at
             # this abstract level, but so far I think all image formats we know
             # do need 4,4.
-            # Copy affine to  isolate from environment
-            affine = np.array(affine, copy=True)
+            # Copy affine to  isolate from environment.  Specify float type to
+            # avoid surprising integer rounding when setting values into affine
+            affine = np.array(affine, dtype=np.float64, copy=True)
             if not affine.shape == (4,4):
                 raise ValueError('Affine should be shape 4,4')
         self._affine = affine
